DataViews Extensibility: Allow unregistering trash post action
